Separated CI vm's parallel methods to a module and fixed an invalid method name.
