+++ integration_test/lib/vnspec.rb
@@ -16,6 +16,7 @@ require 'pry'
 require_relative 'ext/hash'
 require_relative 'ext/erb'
 require_relative 'vnspec/logger'
+require_relative 'vnspec/parallel_module'
 require_relative 'vnspec/config'
 require_relative 'vnspec/ssh'
 require_relative 'vnspec/invoker'
